Conflicts:
	lib/gamelib/gtime.c
	lib/netplay/netplay.c
	macosx/Warzone.xcodeproj/project.pbxproj
	src/multiint.c
	src/multisync.c

+++ src/move.h
@@ -32,7 +32,7 @@ extern "C"
 #endif //__cplusplus
 
 /* The base movement speed */
-extern float	baseSpeed;
+extern int	baseSpeed;
 
 // The next object that should get the router when a lot of units are
 // in a MOVEROUTE state
